+++ packages/gamut-tests/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [2.2.4](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.2.3...@codecademy/gamut-tests@2.2.4) (2021-03-19)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.2.3](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.2.2...@codecademy/gamut-tests@2.2.3) (2021-03-18)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.2.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.2.1...@codecademy/gamut-tests@2.2.2) (2021-03-16)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
 ### [2.2.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.2.0...@codecademy/gamut-tests@2.2.1) (2021-03-04)
 
 **Note:** Version bump only for package @codecademy/gamut-tests